requirements.txt
Django==2.1.3
djangorestframework==3.9.0
pygments==2.2.0
